Raise an error if delivery fails
In the event of an error, Net::HTTP doesn't raise an error - it just
returns a response with a non-200 code. Unless the caller inspects the
response (which is not a common thing to do when sending an email) there
is no sign that anything has gone wrong.

Failing with an error makes these failures more visible.